better balls.

+++ src/Mtl.java
@@ -1,6 +1,6 @@
 // general material class used for all things
 public class Mtl {
-    int[] color;
-    int[] emissionColor;
+    double[] color;
+    double[] emissionColor;
     double emissionStrength;
 }
